Updates to node + moved to search.py
